agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.64
langchain-ollama==0.3.3
langsmith==0.3.45
ollama==0.5.1
orjson==3.10.18
packaging==24.2
pydantic==2.11.5
pydantic_core==2.33.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
zstandard==0.23.0
python-dotenv
